@@ -1,60 +1,57 @@
 {
-  "name": "saturn-datepicker",
-  "version": "1.1.0",
+  "name": "ng-packaged",
+  "version": "0.0.0",
+  "private": true,
   "license": "MIT",
-  "main": "./bundles/saturn-datepicker.umd.js",
-  "module": "./esm5/saturn-datepicker.es5.js",
-  "es2015": "./esm2015/saturn-datepicker.js",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build --prod",
+    "build": "ng build",
+    "build:app": "cpr dist/saturn-datepicker src/saturn-datepicker && ng build packages && rimraf src/@my",
+    "build:lib": "rimraf dist && ng-packagr -p saturn-datepicker/package.json",
     "test": "ng test",
     "lint": "ng lint",
-    "build:lib": "rimraf dist && ng-packagr -p saturn-datepicker/package.json && cp saturn-datepicker/src/*css dist",
-    "packagr": "ng-packagr -p ng-package.json"
+    "e2e": "ng e2e"
   },
-  "private": true,
-  "dependencies": {},
-  "devDependencies": {
-    "@angular-devkit/build-angular": "~0.5.0",
-    "@angular-devkit/schematics": "^0.5.0",
-    "@angular/cdk": "6.0.0",
-    "@angular/cli": "6.0.0",
-    "@angular/common": "6.0.0",
-    "@angular/compiler": "6.0.0",
-    "@angular/compiler-cli": "6.0.0",
-    "@angular/core": "6.0.0",
-    "@angular/forms": "6.0.0",
-    "@angular/http": "6.0.0",
-    "@angular/language-service": "6.0.0",
-    "@angular/material": "6.0.0",
-    "@angular/platform-browser": "6.0.0",
-    "@angular/platform-browser-dynamic": "6.0.0",
-    "@angular/router": "6.0.0",
-    "@types/jasmine": "^2.8.7",
-    "@types/node": "^9.6.12",
-    "codelyzer": "~4.3.0",
+  "dependencies": {
+    "@angular/common": "6.1.0",
+    "@angular/compiler": "6.1.0",
+    "@angular/core": "6.1.0",
+    "@angular/forms": "6.1.0",
+    "@angular/platform-browser": "6.1.0",
+    "@angular/platform-browser-dynamic": "6.1.0",
+    "@angular/animations": "6.1.0",
+    "@angular/material": "^6.4.2",
+    "@angular/cdk": "^6.4.2",
     "core-js": "^2.4.1",
+    "rxjs": "^6.2.2",
+    "zone.js": "0.8.26"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.7.0-rc",
+    "@angular-devkit/schematics": "^0.7.0-rc",
+    "@angular/cli": "6.1.0",
+    "@angular/compiler-cli": "6.1.0",
+    "@angular/language-service": "6.1.0",
+    "@types/jasmine": "~2.8.0",
+    "@types/node": "~10.5.0",
+    "codelyzer": "~4.4.0",
     "cpr": "^3.0.0",
     "jasmine-core": "~3.1.0",
     "jasmine-spec-reporter": "~4.2.0",
     "karma": "~2.0.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.3.0",
+    "karma-coverage-istanbul-reporter": "^2.0.0",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^1.0.0",
-    "ng-packagr": "3.0.0-rc.2",
-    "protractor": "~5.3.0",
+    "ng-packagr": "4.0.0",
+    "protractor": "~5.4.0",
     "rimraf": "^2.6.1",
-    "rxjs": "6.0.0",
-    "rxjs-compat": "^6.1.0",
-    "systemjs": "0.19.43",
-    "ts-node": "^6.0.3",
-    "tsickle": "^0.27.3",
+    "ts-node": "~7.0.0",
+    "tsickle": "^0.32.0",
     "tslib": "^1.7.1",
-    "tslint": "~5.9.0",
+    "tslint": "~5.11.0",
     "typescript": "~2.7.2"
   }
 }
