@@ -10,36 +10,32 @@ import {Directionality} from '@angular/cdk/bidi';
 import {coerceBooleanProperty} from '@angular/cdk/coercion';
 import {ESCAPE, UP_ARROW} from '@angular/cdk/keycodes';
 import {
-    FlexibleConnectedPositionStrategy,
-    Overlay,
-    OverlayConfig,
-    OverlayRef,
-    PositionStrategy,
-    ScrollStrategy,
+  Overlay,
+  OverlayConfig,
+  OverlayRef,
+  PositionStrategy,
+  ScrollStrategy,
 } from '@angular/cdk/overlay';
 import {ComponentPortal, ComponentType} from '@angular/cdk/portal';
 import {DOCUMENT} from '@angular/common';
-import {filter, take} from 'rxjs/operators';
+import {take, filter} from 'rxjs/operators';
 import {
-    AfterViewInit,
-    ChangeDetectionStrategy,
-    ChangeDetectorRef,
-    Component,
-    ComponentRef,
-    ElementRef,
-    EventEmitter,
-    Inject,
-    inject,
-    InjectionToken,
-    Input,
-    NgZone,
-    OnDestroy,
-    OnInit,
-    Optional,
-    Output,
-    ViewChild,
-    ViewContainerRef,
-    ViewEncapsulation,
+  AfterViewInit,
+  ChangeDetectionStrategy,
+  Component,
+  ComponentRef,
+  ElementRef,
+  EventEmitter,
+  Inject,
+  InjectionToken,
+  Input,
+  NgZone,
+  Optional,
+  Output,
+  ViewChild,
+  ViewContainerRef,
+  ViewEncapsulation,
+  OnDestroy,
 } from '@angular/core';
 import {CanColor, mixinColor, ThemePalette} from '@angular/material/core';
 import {MatDialog, MatDialogRef} from '@angular/material/dialog';
@@ -55,16 +51,19 @@ let datepickerUid = 0;
 
 /** Injection token that determines the scroll handling while the calendar is open. */
 export const MAT_DATEPICKER_SCROLL_STRATEGY =
-    new InjectionToken<() => ScrollStrategy>('sat-datepicker-scroll-strategy', {
-      providedIn: 'root',
-      factory: MAT_DATEPICKER_SCROLL_STRATEGY_FACTORY,
-    });
+    new InjectionToken<() => ScrollStrategy>('sat-datepicker-scroll-strategy');
 
 /** @docs-private */
-export function MAT_DATEPICKER_SCROLL_STRATEGY_FACTORY(): () => ScrollStrategy {
-  const overlay = inject(Overlay);
+export function MAT_DATEPICKER_SCROLL_STRATEGY_FACTORY(overlay: Overlay): () => ScrollStrategy {
   return () => overlay.scrollStrategies.reposition();
 }
+
+/** @docs-private */
+export const MAT_DATEPICKER_SCROLL_STRATEGY_FACTORY_PROVIDER = {
+  provide: MAT_DATEPICKER_SCROLL_STRATEGY,
+  deps: [Overlay],
+  useFactory: MAT_DATEPICKER_SCROLL_STRATEGY_FACTORY,
+};
 
 // Boilerplate for applying mixins to SatDatepickerContent.
 /** @docs-private */
@@ -89,7 +88,6 @@ export const _SatDatepickerContentMixinBase = mixinColor(SatDatepickerContentBas
     'class': 'mat-datepicker-content',
     '[@transformPanel]': '"enter"',
     '[class.mat-datepicker-content-touch]': 'datepicker.touchUi',
-    '[class.mat-datepicker-content-above]': '_isAbove',
   },
   animations: [
     matDatepickerAnimations.transformPanel,
@@ -101,10 +99,7 @@ export const _SatDatepickerContentMixinBase = mixinColor(SatDatepickerContentBas
   inputs: ['color'],
 })
 export class SatDatepickerContent<D> extends _SatDatepickerContentMixinBase
-  implements AfterViewInit, CanColor, OnInit, OnDestroy {
-
-  /** Subscription to changes in the overlay's position. */
-  private _positionChange: Subscription|null;
+  implements AfterViewInit, CanColor {
 
   /** Reference to the internal calendar component. */
   @ViewChild(SatCalendar) _calendar: SatCalendar<D>;
@@ -115,42 +110,12 @@ export class SatDatepickerContent<D> extends _SatDatepickerContentMixinBase
   /** Whether the datepicker is above or below the input. */
   _isAbove: boolean;
 
-  constructor(
-    elementRef: ElementRef,
-    private _changeDetectorRef: ChangeDetectorRef,
-    private _ngZone: NgZone) {
+  constructor(elementRef: ElementRef) {
     super(elementRef);
-  }
-
-  ngOnInit() {
-    if (!this.datepicker._popupRef || this._positionChange) {
-      return;
-    }
-
-    const positionStrategy =
-      this.datepicker._popupRef.getConfig().positionStrategy! as FlexibleConnectedPositionStrategy;
-
-    this._positionChange = positionStrategy.positionChanges.subscribe(change => {
-      const isAbove = change.connectionPair.overlayY === 'bottom';
-
-      if (isAbove !== this._isAbove) {
-        this._ngZone.run(() => {
-          this._isAbove = isAbove;
-          this._changeDetectorRef.markForCheck();
-        });
-      }
-    });
   }
 
   ngAfterViewInit() {
     this._calendar.focusActiveCell();
-  }
-
-  ngOnDestroy() {
-    if (this._positionChange) {
-      this._positionChange.unsubscribe();
-      this._positionChange = null;
-    }
   }
 }
 
@@ -408,7 +373,7 @@ export class SatDatepicker<D> implements OnDestroy, CanColor {
             this.beginDate = this.endDate = this._selected = null;
             return;
           }
-          if (this.rangeMode) {
+          if (value && value.hasOwnProperty('begin') && value.hasOwnProperty('end')) {
             value = <SatDatepickerRangeValue<D>>value;
             if (value.begin && value.end &&
               this._dateAdapter.compareDate(value.begin, value.end) <= 0) {
@@ -483,7 +448,7 @@ export class SatDatepicker<D> implements OnDestroy, CanColor {
   /** Open the calendar as a dialog. */
   private _openAsDialog(): void {
     this._dialogRef = this._dialog.open<SatDatepickerContent<D>>(SatDatepickerContent, {
-      direction: this._getDirection(),
+      direction: this._dir ? this._dir.value : 'ltr',
       viewContainerRef: this._viewContainerRef,
       panelClass: 'mat-datepicker-dialog',
     });
@@ -505,7 +470,6 @@ export class SatDatepicker<D> implements OnDestroy, CanColor {
     }
 
     if (!this._popupRef.hasAttached()) {
-      this._popupRef.setDirection(this._getDirection());
       this._popupComponentRef = this._popupRef.attach(this._calendarPortal);
       this._popupComponentRef.instance.datepicker = this;
       this._setColor();
@@ -523,12 +487,13 @@ export class SatDatepicker<D> implements OnDestroy, CanColor {
       positionStrategy: this._createPopupPositionStrategy(),
       hasBackdrop: true,
       backdropClass: 'mat-overlay-transparent-backdrop',
-      direction: this._getDirection(),
+      direction: this._dir,
       scrollStrategy: this._scrollStrategy(),
       panelClass: 'mat-datepicker-popup',
     });
 
     this._popupRef = this._overlay.create(overlayConfig);
+    this._popupRef.overlayElement.setAttribute('role', 'dialog');
 
     merge(
       this._popupRef.backdropClick(),
@@ -544,7 +509,8 @@ export class SatDatepicker<D> implements OnDestroy, CanColor {
   /** Create the popup PositionStrategy. */
   private _createPopupPositionStrategy(): PositionStrategy {
     return this._overlay.position()
-      .flexibleConnectedTo(this._datepickerInput.getPopupConnectionElementRef())
+      .flexibleConnectedTo(this._datepickerInput.getConnectedOverlayOrigin())
+      .withTransformOriginOn('.mat-datepicker-content')
       .withFlexibleDimensions(false)
       .withViewportMargin(8)
       .withPush(false)
@@ -593,10 +559,5 @@ export class SatDatepicker<D> implements OnDestroy, CanColor {
     if (this._dialogRef) {
       this._dialogRef.componentInstance.color = color;
     }
-  }
-
-  /** Returns the layout direction of the datepicker. */
-  private _getDirection() {
-    return this._dir ? this._dir.value : 'ltr';
   }
 }
