@@ -9,7 +9,7 @@
 /** @docs-private */
 export function createMissingDateImplError(provider: string) {
     return Error(
-        `MatDatepicker: No provider found for ${provider}. You must import one of the following ` +
-        `modules at your application root: MatNativeDateModule, MatMomentDateModule, or provide a ` +
-        `custom implementation.`)
+        `SatDatepicker: No provider found for ${provider}. You must import one of the following ` +
+        `modules at your application root: SatNativeDateModule, MatMomentDateModule, or provide a ` +
+        `custom implementation.`);
 }
