@@ -10,15 +10,15 @@ import {InjectionToken} from '@angular/core';
 
 
 export type MatDateFormats = {
-    parse: {
-        dateInput: any
-    },
-    display: {
-        dateInput: any,
-        monthYearLabel: any,
-        dateA11yLabel: any,
-        monthYearA11yLabel: any,
-    }
+  parse: {
+    dateInput: any
+  },
+  display: {
+    dateInput: any,
+    monthYearLabel: any,
+    dateA11yLabel: any,
+    monthYearA11yLabel: any,
+  }
 };
 
 
